Introduce internal and spi packages
