+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- **Breaking** Introduced `spi` and `internal` packages to clearly
+specify user API
 - **Breaking** Matcher properties with names starting with `_` are no longer bound to the
 pattern match
