+++ src/internal/MicrogrammarUpdates.ts
@@ -1,5 +1,6 @@
 import { ChangeSet } from "./ChangeSet";
-import { isTreePatternMatch, PatternMatch } from "./PatternMatch";
+
+import { isTreePatternMatch, PatternMatch } from "../PatternMatch";
 
 export interface MatchUpdater {
     newContent(): string;
