+++ src/spi/StringInputStream.ts
